Move ring testing stuff in testing module
